+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add support for OpenSearch database type
 - Add `database index list` and `database index` commands for managing OpenSearch database indices 
+- Add `--label` flag to `server create` and `server modify` commands
 
 ## [2.7.1] - 2023-05-16
 ### Fixed